Add files via upload
ESP32 I2C ISSUE SOLVED:
Version 1.0.1
The FDC1004 is a single-chip IC for capacitance measurement for applications including proximity sensing and liquid level sensing. This is based on the concept of measuring the capacitance of an electrode with respect to the ground.

This breakout board contains everything to connect it to an Arduino's I2C pins, and it can be connected to any other platform with support for I2C two-wire interfaces.